removed old fragments, repositories, dis, usecases
