--- app/src/main/java/com/themovieviewer/di/RepositoryModule.kt
@@ -1,9 +1,0 @@
-package com.themovieviewer.di
-
-import dagger.Module
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-
-@Module
-@InstallIn(SingletonComponent::class)
-object RepositoryModule
