--- app/src/main/java/com/themovieviewer/util/Constants.kt
@@ -1,4 +1,0 @@
-package com.themovieviewer.util
-
-const val TAG = "AppDebug"
-const val DATABASE_NAME = "movie_viewer.db"
