--- app/src/main/java/com/themovieviewer/util/Executores.kt
@@ -1,9 +1,0 @@
-package com.themovieviewer.util
-
-import java.util.concurrent.Executors
-
-private val IO_EXECUTOR = Executors.newSingleThreadExecutor()
-
-fun ioThread(f: () -> Unit) {
-    IO_EXECUTOR.execute(f)
-}
