--- app/src/main/res/layout/widget_credits_vertical.xml
@@ -1,73 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<layout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    xmlns:app="http://schemas.android.com/apk/res-auto">
-    <data>
-        <variable
-            name="model"
-            type="com.themovieviewer.core.data.network.model.CreditsCastCrewDto" />
-    </data>
-    <androidx.constraintlayout.widget.ConstraintLayout
-        android:background="@android:color/white"
-        android:padding="5dp"
-        android:layout_width="135dp"
-        android:layout_height="250dp">
-
-        <androidx.cardview.widget.CardView
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            app:cardCornerRadius="4dp"
-            app:cardUseCompatPadding="true"
-            app:cardBackgroundColor="@android:color/white"
-            app:layout_constraintTop_toTopOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintEnd_toEndOf="parent">
-
-            <androidx.constraintlayout.widget.ConstraintLayout
-                android:layout_width="match_parent"
-                android:layout_height="match_parent"
-                app:layout_constraintTop_toTopOf="parent"
-                app:layout_constraintBottom_toBottomOf="parent"
-                app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintEnd_toEndOf="parent">
-
-                <ImageView
-                    android:id="@+id/profile"
-                    android:scaleType="fitXY"
-                    android:layout_width="0dp"
-                    android:layout_height="135dp"
-                    tools:src="@drawable/test"
-                    app:loadImage="@{model.profile_path}"
-                    app:layout_constraintTop_toTopOf="parent"
-                    app:layout_constraintStart_toStartOf="parent"
-                    app:layout_constraintEnd_toEndOf="parent"/>
-
-                <TextView
-                    android:padding="5dp"
-                    android:id="@+id/originalName"
-                    android:layout_width="0dp"
-                    android:layout_height="wrap_content"
-                    android:textColor="@android:color/black"
-                    android:textStyle="bold"
-                    android:textSize="15sp"
-                    android:text="@{model.original_name}"
-                    tools:text="aaaaaaaa"
-                    app:layout_constraintTop_toBottomOf="@+id/profile"
-                    app:layout_constraintStart_toStartOf="parent"
-                    app:layout_constraintEnd_toEndOf="parent"/>
-
-                <TextView
-                    android:padding="5dp"
-                    android:id="@+id/character"
-                    android:layout_width="0dp"
-                    android:layout_height="wrap_content"
-                    android:textColor="@android:color/black"
-                    android:text="@{model.character}"
-                    tools:text="character"
-                    app:layout_constraintTop_toBottomOf="@+id/originalName"
-                    app:layout_constraintStart_toStartOf="parent"
-                    app:layout_constraintEnd_toEndOf="parent"/>
-            </androidx.constraintlayout.widget.ConstraintLayout>
-        </androidx.cardview.widget.CardView>
-    </androidx.constraintlayout.widget.ConstraintLayout>
-</layout>
