--- app/src/main/res/layout/widget_people_acting.xml
@@ -1,65 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<layout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    xmlns:app="http://schemas.android.com/apk/res-auto">
-    <data>
-        <variable
-            name="model"
-            type="com.themovieviewer.core.model.data.CastCrew" />
-    </data>
-    <androidx.constraintlayout.widget.ConstraintLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content">
-        <androidx.constraintlayout.widget.Guideline
-            android:id="@+id/startGuideLine"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:orientation="vertical"
-            app:layout_constraintGuide_percent="0.05"/>
-
-        <TextView
-            android:id="@+id/tv_releaseDate"
-            android:layout_width="40dp"
-            android:layout_height="wrap_content"
-            android:textColor="@android:color/black"
-            android:gravity="center"
-            app:release_date="@{model.release_date}"
-            app:layout_constraintTop_toTopOf="parent"
-            app:layout_constraintStart_toStartOf="@+id/startGuideLine"
-            tools:text="2022"/>
-
-        <TextView
-            android:id="@+id/tv_title"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginStart="10dp"
-            android:text="@{model.title}"
-            android:textColor="@android:color/black"
-            app:layout_constraintTop_toTopOf="parent"
-            app:layout_constraintStart_toEndOf="@+id/tv_releaseDate"
-            tools:text="Cruella"/>
-
-        <TextView
-            android:id="@+id/tv_role"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:textColor="@android:color/darker_gray"
-            android:layout_marginStart="5dp"
-            android:text="@string/tv_as"
-            app:layout_constraintTop_toTopOf="parent"
-            app:layout_constraintStart_toEndOf="@+id/tv_title"
-            tools:text="as"/>
-
-        <TextView
-            android:id="@+id/tv_character"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:textColor="@android:color/darker_gray"
-            android:textStyle="bold"
-            android:text="@{model.character}"
-            android:layout_marginStart="5dp"
-            app:layout_constraintTop_toTopOf="parent"
-            app:layout_constraintStart_toEndOf="@+id/tv_role"
-            tools:text="Estella / Cruella"/>
-    </androidx.constraintlayout.widget.ConstraintLayout>
-</layout>
