--- app/src/main/res/menu/activity_main_drawer.xml
@@ -1,16 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<menu xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    tools:showIn="navigation_view">
-
-    <group android:checkableBehavior="single">
-        <item
-            android:id="@+id/nav_main"
-            android:icon="@drawable/ic_menu_camera"
-            android:title="@string/menu_main" />
-        <item
-            android:id="@+id/nav_tv_show"
-            android:icon="@drawable/ic_menu_camera"
-            android:title="@string/menu_tv_show" />
-    </group>
-</menu>
