--- app/src/main/res/menu/movie.xml
@@ -1,25 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<menu xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto">
-    <item
-        android:id="@+id/action_add"
-        android:orderInCategory="100"
-        android:title="@string/action_add"
-        android:icon="@drawable/ic_favorite_24"
-        android:visible="false"
-        app:showAsAction="ifRoom" />
-    <item
-        android:id="@+id/action_remove"
-        android:orderInCategory="100"
-        android:title="@string/action_remove"
-        android:visible="false"
-        android:icon="@drawable/ic_remove_24"
-        app:showAsAction="ifRoom" />
-    <item
-        android:id="@+id/action_done"
-        android:orderInCategory="100"
-        android:title="@string/action_done"
-        android:icon="@drawable/ic_done_24"
-        android:visible="false"
-        app:showAsAction="ifRoom" />
-</menu>
