--- app/src/main/res/navigation/mobile_navigation.xml
@@ -1,55 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<navigation xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/mobile_navigation"
-    app:startDestination="@+id/nav_main">
-
-    <fragment
-        android:id="@+id/nav_main"
-        android:name="com.themovieviewer.presentation.ui.main.MainFragment"
-        android:label="@string/menu_main"
-        tools:layout="@layout/fragment_main" >
-    </fragment>
-
-    <fragment
-        android:id="@+id/nav_people_details"
-        android:name="com.themovieviewer.presentation.ui.people.PeopleDetailsFragment"
-        android:label="@string/menu_people_details"
-        tools:layout="@layout/fragment_people" >
-        <argument
-            android:name="personId"
-            app:argType="integer" />
-        <action
-            android:id="@+id/action_people_details_to_movie_details"
-            app:destination="@+id/nav_movie_details" />
-    </fragment>
-
-    <fragment
-        android:id="@+id/nav_top_rated"
-        android:name="com.themovieviewer.presentation.ui.toprated.TopRatedFragment"
-        android:label="@string/menu_tv_show"
-        tools:layout="@layout/fragment_top_rated">
-        <action
-            android:id="@+id/action_TopRatedFragment_to_MovieDetailsFragment"
-            app:destination="@+id/nav_movie_details" />
-    </fragment>
-
-    <fragment
-        android:id="@+id/nav_now_playing"
-        android:name="com.themovieviewer.presentation.ui.nowplaying.NowPlayingFragment"
-        android:label="@string/menu_gallery"
-        tools:layout="@layout/fragment_now_palying">
-        <action
-            android:id="@+id/action_NowPlayingFragment_to_MovieDetailsFragment"
-            app:destination="@+id/nav_movie_details" />
-    </fragment>
-    <fragment
-        android:id="@+id/nav_popular"
-        android:name="com.themovieviewer.presentation.ui.popular.MoviePopularFragment"
-        android:label="@string/menu_Popular"
-        tools:layout="@layout/fragment_popular" >
-        <action
-            android:id="@+id/action_MoviePopularFragment_to_MovieDetailsFragment"
-            app:destination="@+id/nav_movie_details" />
-    </fragment></navigation>
