+++ core-data/src/main/java/com/themovieviewer/core/data/di/DataModule.kt
@@ -1,0 +1,28 @@
+package com.themovieviewer.core.data.di
+
+import com.google.gson.GsonBuilder
+import com.themovieviewer.core.data.network.MovieService
+import com.themovieviewer.core.data.network.datasource.NowPlayingDataSource
+import com.themovieviewer.core.data.repository.MovieRepository
+import dagger.Module
+import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+import retrofit2.Retrofit
+import retrofit2.converter.gson.GsonConverterFactory
+import javax.inject.Singleton
+
+
+@Module
+@InstallIn(SingletonComponent::class)
+object DataModule {
+    @Singleton
+    @Provides
+    fun provideMovieService(): MovieService {
+        return Retrofit.Builder()
+            .baseUrl("https://api.themoviedb.org/3/")
+            .addConverterFactory(GsonConverterFactory.create(GsonBuilder().create()))
+            .build()
+            .create(MovieService::class.java)
+    }
+}
